First pass at catching errors
Using a bounds check, and passing around Result types.

If the bounds check fails, the original value is returned as an Err
Because the check function is called from inside a try!(), this causes
an early return of another Result. The calling function checks for this,
inserting a -1 value in that case.

check() <-- Return Result
    |_convert_bng() <-- early return in case of Err Result
                |_convert_to_bng_threaded_vec() <-- check for Err & insert -1