Added some comments to the interpolation function
It should be possible to switch this function to using SIMD.
SSE 4.1 has a dot product function, but it isn't clear whether this
is the best route to take. Alternative approaches involve splitting
the grid intersections into vectors of x, y, and z values (eg an SoA
approach) and operating on those instead. See e.g.
http://fastcpp.blogspot.com/2011/06/bilinear-pixel-interpolation-using-sse.html

[ci skip]